SRSUE: Fixed CLang Tidy in ue.cc

+++ lib/include/srslte/phy/common/phy_common.h
@@ -352,7 +352,7 @@ SRSLTE_API char *srslte_mod_string(srslte_mod_t mod);
 
 SRSLTE_API uint32_t srslte_mod_bits_x_symbol(srslte_mod_t mod);
 
-SRSLTE_API int srslte_band_get_band(uint32_t dl_earfcn);
+SRSLTE_API uint8_t srslte_band_get_band(uint32_t dl_earfcn);
 
 SRSLTE_API bool srslte_band_is_tdd(uint32_t band);
